Algunas funcionalidades (sin probar) añadidas
Con respecto a OrderDetails, se avanzó en cuanto a las funcionalidades del Service. Sin embargo, aun faltan implementarse en el controller, y las actuales faltan probarlas

+++ FoodYeah-OpenSource/FoodYeah/src/main/java/com/example/entity/OrderDetail.java
@@ -39,10 +39,10 @@ public class OrderDetail {
     @JsonIgnoreProperties({"hibernateLazyInitializer","handler"})
     private Product product;
 
-    @ApiModelProperty(value="La cantidad de pedidos del producto", dataType="byte", position=4)
+    @ApiModelProperty(value="La cantidad de pedidos del producto", dataType="int", position=4)
     @NotNull(message = "La cantidad no puede estar vacía")
     @Column(name = "quantity", nullable = false)
-    private byte quantity;
+    private int quantity;
 
     @ApiModelProperty(value="El precio unitario", dataType="float", position=5)
     @NotNull(message = "El precio unitario no puede ser vacío")
